Yet another build fix for Windows. https://webkit.org/b/142513>
Reviewed by Alex Christensen.

Looks like MSVC requires the function be explicitly declared in a header file
in order for it to be linkable from another file in the same project.  This is
strange, but it seems to make MSVC happy.

Also fixed a typo in testapi.vcxproj.filters.

* API/tests/CompareAndSwapTest.cpp:
* API/tests/CompareAndSwapTest.h: Added.
* API/tests/testapi.c:
* JavaScriptCore.vcxproj/testapi/testapi.vcxproj:
* JavaScriptCore.vcxproj/testapi/testapi.vcxproj.filters:
* JavaScriptCore.xcodeproj/project.pbxproj:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@181315 268f45cc-cd09-0410-ab3c-d52691b4dbfc

Conflicts:
	Source/JavaScriptCore/ChangeLog

Yet another build fix for Windows.