Object allocation sinking phase shouldn't re-decorate previously sunken allocations on each fixpoint operation https://bugs.webkit.org/show_bug.cgi?id=142686
Reviewed by Oliver Hunt.

Just because promoteHeapAccess() notifies us of an effect to a heap location in a node doesn't
mean that we should handle it as if it was for one of our sinking candidates. Instead we should
prune based on m_sinkCandidates.

This fixes a benign bug where we would generate a lot of repeated IR for some pathological
tests.

* dfg/DFGObjectAllocationSinkingPhase.cpp:
(JSC::DFG::ObjectAllocationSinkingPhase::promoteSunkenFields):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@181495 268f45cc-cd09-0410-ab3c-d52691b4dbfc

Conflicts:
	Source/JavaScriptCore/ChangeLog

Object allocation sinking phase shouldn't re-decorate previously sunken allocations on each fixpoint operation