In the 64-bit DFG and FTL, Array::Double case for HasIndexedProperty should set its result to true when all is well. <https://webkit.org/b/143396>
Reviewed by Filip Pizlo.

The DFG was neglecting to set the result boolean.  The FTL was setting it with
an inverted value.  Both of these are now resolved.

* dfg/DFGSpeculativeJIT64.cpp:
(JSC::DFG::SpeculativeJIT::compile):
* ftl/FTLLowerDFGToLLVM.cpp:
(JSC::FTL::LowerDFGToLLVM::compileHasIndexedProperty):
* tests/stress/for-in-array-mode.js: Added.
(.):
(test):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@182444 268f45cc-cd09-0410-ab3c-d52691b4dbfc

Conflicts:
	Source/JavaScriptCore/ChangeLog

In the 64-bit DFG and FTL, Array::Double case for HasIndexedProperty should set its result to true when all is well.